backend/requirements.txt
langgraph
requests
python-dotenv
fastapi
uvicorn
sqlalchemy
psycopg2-binary
supabase
pydantic
